requirements.txt
fastapi
streamlit
python-multipart
pandas
Pillow
uvicorn
mecab-python3
unidic-lite